refactor: moved util to internal
