+++ internal/util/removeBadCharacters.go
