+++ internal/util/timestampToFloat.go
